+++ changes.md
@@ -9,7 +9,7 @@
 3. Provide full support of require.js. One could load now JSRootCore.js script like:
       <script type="text/javascript" src="require.js" data-main="scripts/JSRootCore.js"></script>
    After this several modules are defined and can be used with syntax like:
-      require(['JSRootIO'], function(JSROOT) { /*any user code*/});
+      require(['JSRootPainter'], function(jsroot) { /*any user code*/});
    Also inside JSROOT require.js used to load all dependencies.